+++ src/app/components/about/about.ts
