+++ src/app/components/stock/stock.html
