+++ src/app/models/provent.model.ts
@@ -1,0 +1,10 @@
+import { Stock } from './stock.model';
+
+export interface Provent {
+  id: number;
+  date: string;
+  type: string;
+  stock: Stock;
+  valueTotal: number;
+  quantity: number;
+}
