+++ src/app/models/resume.model.ts
@@ -1,0 +1,8 @@
+export interface Resume {
+  stockCount: number;
+  totalInvested: number;
+  averagePrice: number;
+  totalProvents: number;
+  totalProventsLast12Months: number;
+  averageProventsLast4Months: number;
+}
