+++ src/app/models/stock.model.ts
@@ -1,0 +1,4 @@
+export interface Stock {
+  tick: string;
+  company: string;
+}
